+++ structured_derby/src/main/java/com/splicemachine/derby/impl/sql/execute/LocalWriteContextFactory.java
@@ -1,5 +1,6 @@
 package com.splicemachine.derby.impl.sql.execute;
 
+import com.carrotsearch.hppc.BitSet;
 import com.splicemachine.constants.SpliceConstants;
 import com.splicemachine.derby.ddl.DDLChange;
 import com.splicemachine.derby.ddl.DDLCoordinationFactory;
@@ -15,6 +16,7 @@ import com.splicemachine.si.impl.DDLFilter;
 import com.splicemachine.si.impl.TransactionId;
 import com.splicemachine.tools.ResettableCountDownLatch;
 import com.splicemachine.utils.SpliceLogUtils;
+
 import org.apache.derby.catalog.IndexDescriptor;
 import org.apache.derby.catalog.UUID;
 import org.apache.derby.iapi.error.StandardException;
@@ -28,7 +30,6 @@ import org.apache.log4j.Logger;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.sql.SQLException;
-import java.util.BitSet;
 import java.util.Set;
 import java.util.concurrent.CopyOnWriteArraySet;
 import java.util.concurrent.TimeUnit;
@@ -454,7 +455,7 @@ public class LocalWriteContextFactory implements WriteContextFactory<RegionCopro
         }
 
         private static int[] getMainColToIndexPosMap(int[] indexColsToMainColMap, BitSet indexedCols) {
-            int[] mainColToIndexPosMap = new int[indexedCols.length()];
+            int[] mainColToIndexPosMap = new int[(int)indexedCols.length()];
             for(int indexCol=0;indexCol<indexColsToMainColMap.length;indexCol++){
                 int mainCol = indexColsToMainColMap[indexCol];
                 mainColToIndexPosMap[mainCol-1] = indexCol;
