+++ structured_derby/src/main/java/com/splicemachine/derby/impl/sql/execute/index/IndexTransformer.java
@@ -1,5 +1,6 @@
 package com.splicemachine.derby.impl.sql.execute.index;
 
+import com.carrotsearch.hppc.BitSet;
 import com.splicemachine.derby.hbase.SpliceDriver;
 import com.splicemachine.encoding.Encoding;
 import com.splicemachine.encoding.MultiFieldDecoder;
@@ -8,10 +9,8 @@ import com.splicemachine.storage.EntryAccumulator;
 import com.splicemachine.storage.EntryDecoder;
 import com.splicemachine.storage.SparseEntryAccumulator;
 import com.splicemachine.storage.index.BitIndex;
-
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.util.BitSet;
 
 /**
  * Class responsible for transforming an incoming main table row
@@ -97,7 +96,7 @@ public class IndexTransformer {
             }
         }
         //add the row location to the end of the index row
-        indexRowAccumulator.add(translatedIndexedColumns.length(),ByteBuffer.wrap(Encoding.encodeBytesUnsorted(mutation.getRow())));
+        indexRowAccumulator.add((int)translatedIndexedColumns.length(),ByteBuffer.wrap(Encoding.encodeBytesUnsorted(mutation.getRow())));
         byte[] indexRowKey = getIndexRowKey(mutation.getRow());
         byte[] indexRowData = indexRowAccumulator.finish();
         return new KVPair(indexRowKey,indexRowData,mutation.getType());
@@ -117,7 +116,7 @@ public class IndexTransformer {
     public byte[] getIndexRowKey(byte[] rowKey){
         if(isUnique) return indexKeyAccumulator.finish();
         else{
-            indexKeyAccumulator.add(translatedIndexedColumns.length(),ByteBuffer.wrap(rowKey));
+            indexKeyAccumulator.add((int)translatedIndexedColumns.length(),ByteBuffer.wrap(rowKey));
             return indexKeyAccumulator.finish();
         }
     }
